requirements-dev.txt
pytest>=7.1.2
pytest-benchmark>=3.4.1
kytea>=0.1.7
SudachiPy>=0.6.3
SudachiDict-core>=20220519
